Fix handling of globals to ensure persistence and class-wide consistency.
`.cacheCommonInfo` is now responsible for verifying slots affected by global
parameters. It does so by checking if the first encountered instance of a class
already has the affected slots set (i.e., non-NA); if so, it caches that value,
and if not, it gets the global value.  Whatever is picked is then cached in the
SE and used to enfroce the same value across all instances of the class. This
preserves class-wide consistency in a global-responsive manner while respecting
existing non-NA values, e.g., for serialized memory.